#17 Put the components in separate files
