+++ app/src/main/java/com/example/triviaapp/game/component/SeperationLine.kt
@@ -1,0 +1,35 @@
+package com.example.triviaapp.game.component
+
+import androidx.compose.foundation.Canvas
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.PathEffect
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun SeperationLine(pathEffect: PathEffect) {
+    Canvas(
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(1.dp)
+
+    ) {
+        drawLine(
+            color = Color.LightGray,
+            start = Offset(0f, 0f),
+            end = Offset(size.width, 0f),
+            pathEffect = pathEffect
+        )
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun SeperationLinePreview() {
+    SeperationLine(pathEffect = PathEffect.dashPathEffect(floatArrayOf(10f, 10f), phase = 0f))
+}
